refactor: PMI 함수를 soynlp.word에서 soynlp.utils로 이동 (#267)
- soynlp/utils/pmi.py: PMI 관련 함수(_as_diag, _logarithm_and_ppmi, pmi) 이동
- soynlp/utils/__init__.py: pmi를 public export에 추가
- soynlp/word/pmi.py: soynlp.utils.pmi를 deprecated re-export로 교체
  (DeprecationWarning 포함, 하위 호환 유지)
- tests/unit/test_pmi.py: soynlp.utils.pmi 기준으로 import 변경,
  deprecated 경로 경고 검증 테스트 추가
